perf(launch): a new market is listed with its logo and priced immediately
Two waits stood between creating a market and it being fully live. Neither was
inherent — both were defaults sized for a world that no longer applies.

1. The logo took up to 10 minutes

The wizard never sends logo_url, so a new market was inserted with null and only
gained its real image when the indexer's metadata pass next ran. Until then it
rendered as the initials placeholder.

POST /api/markets now resolves the logo inline, before the insert, whenever the
caller supplied none and a mainnet CA is available — so the image is already on
the row the markets page reads on its very first render.

The resolver is not new: app/api/token-logo/[mint] already did exactly this
(GeckoTerminal, then DexScreener for pump.fun-style tokens). It is extracted to
lib/token-logo.ts and both callers now share the one implementation. It swallows
its own errors and a null simply restores the old placeholder behaviour — a logo
is cosmetic and must never fail a launch.

Verified against three real CAs (ANSEM, Jimothy, SOL) — all resolve.

2. The keeper took up to 30 seconds

Poll interval 30s -> 3s, and the feed it polls is now no-store.

The cache was the worse half: /api/playground/registered-markets is a liveness
feed, not a page, and s-maxage=10 with stale-while-revalidate=30 meant the keeper
could be served the pre-launch answer while it refreshed. Combined, a market
could be listed and tradeable in the UI while its price sat at the seed value for
half a minute.

The old interval was sized for a cached Blob-only read. The feed is now one
indexed select over a table with a handful of rows, so 3s is cheap.

tsc clean, build clean, suite unchanged at the pre-existing 4 files / 14 tests.
